Removing changes to types
